requirements.txt
faiss_cpu==1.10.0
mcp==1.9.1
nest_asyncio==1.6.0
numpy==2.2.6
python-dotenv==1.1.0
Requests==2.32.3
starlette==0.46.2
uvicorn==0.34.2
